refactor: move target_response to root
