+++ src/agenteval/targets/bedrock_agent/target.py
@@ -1,6 +1,7 @@
 import uuid
 
-from agenteval.targets import Boto3Target, TargetResponse
+from agenteval.target_response import TargetResponse
+from agenteval.targets import Boto3Target
 
 _SERVICE_NAME = "bedrock-agent-runtime"
